rearrange coretypes
