+++ src/imgtools/coretypes/__init__.py
@@ -1,4 +1,4 @@
 from .box import BoxPadMethod, RegionBox
-from .spatial_types import Coordinate3D, Size3D, Spacing3D
+from .spatial_types.coord_types import Coordinate3D, Size3D, Spacing3D
 
 __all__ = ["Coordinate3D", "Size3D", "Spacing3D", "RegionBox", "BoxPadMethod"]
