+++ src/imgtools/coretypes/spatial_types/README.md
