+++ src/imgtools/coretypes/spatial_types/__init__.py
@@ -1,0 +1,5 @@
+from .coord_types import Coordinate3D, Size3D, Spacing3D
+from .direction import Direction
+from .image_geometry import ImageGeometry
+
+__all__ = ["Coordinate3D", "Spacing3D", "Size3D", "Direction", "ImageGeometry"]
