+++ src/imgtools/coretypes/spatial_types/direction.py
